opea-comps/mega-service/requirements.txt
opea-comps
fastapi
uvicorn
aiohttp